# Employees Can Create GIFs -> Upload Image using route & multer -> Upload to cloudinary and get image URL -> Insert Cloudinary Image URL into DB
